Rename 'logging.py' to 'logger.py' due to name class with the standard module. Make value parsing from excel sheets more robust.
