+++ antibodies/deprecated/instance_analysis_workflow2.py
@@ -17,7 +17,7 @@ from batchlib.segmentation import SeededWatershed
 from batchlib.segmentation.stardist_prediction import StardistPrediction
 from batchlib.segmentation.torch_prediction import TorchPrediction
 from batchlib.segmentation.unet import UNet2D
-from batchlib.util.logging import get_logger
+from batchlib.util.logger import get_logger
 
 logger = get_logger('Workflow.InstanceAnalysis2')
